requirements.txt
aiohttp==3.5.4
discord.py==1.0.1
googletrans==2.3.0
humanize==0.5.1
iso8601==0.1.12
lxml==4.2.3
nltk==3.4
Pillow==6.0.0
psutil==5.4.6
pytesseract==0.2.6
PyYAML==5.1
strconv==0.4.2
typing==3.6.6
Wand==0.4.4
youtube_dl==2019.1.2
